requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
aiohttp==3.9.1
requests>=2.28.0
dingtalk-stream>=0.8.0
python-dotenv>=1.0.0
pycryptodome>=3.18.0
